Move metadata document paths to separate file
