+++ context/app/static/js/components/entity-search/SearchWrapper/metadataDocumentPaths.js
@@ -1,0 +1,19 @@
+const donorMetadataPath = 'mapped_metadata';
+const sampleMetdataPath = 'metadata';
+
+const paths = {
+  donor: {
+    donor: donorMetadataPath,
+  },
+  sample: {
+    sample: sampleMetdataPath,
+    donor: `donor.${donorMetadataPath}`,
+  },
+  dataset: {
+    donor: `donor.${donorMetadataPath}`,
+    sample: `source_sample.${sampleMetdataPath}`,
+    dataset: 'metadata.metadata',
+  },
+};
+
+export { paths };
